Consolidate status deletion logic into dedicated action
