+++ app/javascript/mastodon/containers/status_container.jsx
@@ -2,9 +2,6 @@ import { injectIntl, defineMessages } from 'react-intl';
 
 import { connect } from 'react-redux';
 
-const messages = defineMessages({
-  deleteSuccess: { id: 'status.delete.success', defaultMessage: 'Post deleted.' },
-});
 
 import {
   unmuteAccount,
@@ -38,14 +35,13 @@ import { initReport } from '../actions/reports';
 import {
   muteStatus,
   unmuteStatus,
-  deleteStatus,
   toggleStatusSpoilers,
   toggleStatusCollapse,
   editStatus,
   translateStatus,
   undoStatusTranslation,
 } from '../actions/statuses';
-import { showAlert } from '../actions/alerts';
+import { deleteStatus } from '../actions/delete';
 import { setStatusQuotePolicy } from '../actions/statuses_typed';
 import Status from '../components/status';
 import { deleteModal } from '../initial_state';
@@ -110,21 +106,14 @@ const mapDispatchToProps = (dispatch, { contextType, intl }) => ({
   },
 
   onDelete (status, withRedraft = false) {
-    const handleDeleteSuccess = () => {
-      dispatch(showAlert({
-        message: intl.formatMessage(messages.deleteSuccess),
-      }));
-    };
-
     if (!deleteModal) {
-      dispatch(deleteStatus(status.get('id'), withRedraft)).then(handleDeleteSuccess);
+      dispatch(deleteStatus(status.get('id'), withRedraft));
     } else {
       dispatch(openModal({ 
         modalType: 'CONFIRM_DELETE_STATUS', 
         modalProps: { 
           statusId: status.get('id'), 
-          withRedraft,
-          onDeleteSuccess: handleDeleteSuccess
+          withRedraft
         } 
       }));
     }
