+++ app/javascript/mastodon/locales/en.json
@@ -869,7 +869,7 @@
   "status.continued_thread": "Continued thread",
   "status.copy": "Copy link to post",
   "status.delete": "Delete",
-  "status.delete.success": "Post deleted.",
+  "status.delete.success": "Post deleted",
   "status.detailed_status": "Detailed conversation view",
   "status.direct": "Privately mention @{name}",
   "status.direct_indicator": "Private mention",
